requirements.txt
selenium
openai
python-dotenv
PyGithub
GitPython
